add Toggle and Action support and 'merge' with Release version
this was an attempt to merge the additions to control24osc in release that were done after DEV_otherdevices was branched off and are as a result missing from procontrolosc.py. 

The goal was to restore Toggle support for learn commands (this prevents double triggering of the learned commands) and to add direct support for 'Action': 23453 in the procontrolmap.py where 23453 is Reaper's command ID

Disclaimer: I do not know how to code! this was a mere brute force hack attempt involving much copying and pasting to get it working because the double triggering was pissing me off. But, that said it does appear to be working....